refactor: make grades not climb specific
The plan is to use the types publicized by the `grades.rs` modules as
grades for the entire schema.
